chore(app): organize tests
