+++ core/application/src/tests/mod.rs
@@ -1,0 +1,21 @@
+use lightning_interfaces::partial;
+use lightning_test_utils::json_config::JsonConfigProvider;
+
+use crate::Application;
+
+mod balances;
+mod content_registry;
+mod epoch_change;
+mod everything_else;
+mod genesis;
+mod participation;
+mod pod;
+mod protocol_params;
+mod reputation;
+mod staking;
+mod utils;
+
+partial!(TestBinding {
+    ConfigProviderInterface = JsonConfigProvider;
+    ApplicationInterface = Application<Self>;
+});
